+++ src/helpers/OTPublisherHelper.js
@@ -35,6 +35,31 @@ const sanitizeVideoSource = (videoSource = 'camera') => (videoSource === 'camera
 const sanitizeAudioBitrate = (audioBitrate = 40000) =>
   (audioBitrate < 6000 || audioBitrate > 510000 ? 40000 : audioBitrate);
 
+const sanitizeSubscriberAudioFallback = (audioFallback, audioFallbackEnabled, videoSource) => {
+  if (typeof audioFallback === 'undefined') {
+    if (typeof audioFallbackEnabled !== 'undefined') {
+      return !!audioFallbackEnabled;
+    }
+  }
+  if (typeof audioFallback !== 'object') {
+    return !(videoSource === 'screen');
+  }
+  if (typeof audioFallback.subscriber !== 'undefined') {
+    return !!audioFallback.subscriber;
+  }
+  return !(videoSource === 'screen');
+};
+
+const sanitizePublisherAudioFallback = (audioFallback, videoSource) => {
+  if (typeof audioFallback !== 'object') {
+    return !(videoSource === 'screen');
+  }
+  if (typeof audioFallback.publisher !== 'undefined') {
+    return !!audioFallback.publisher;
+  }
+  return !(videoSource === 'screen');
+};
+
 const sanitizeVideoContentHint = (sanitizeVideoContentHint = '') => {
   switch (sanitizeVideoContentHint) {
     case 'motion':
@@ -58,7 +83,8 @@ const sanitizeProperties = (properties) => {
       publishCaptions: false,
       name: '',
       cameraPosition: 'front',
-      audioFallbackEnabled: true,
+      publisherAudioFallback: false,
+      subscriberAudioFallback: true,
       audioBitrate: 40000,
       enableDtx: false,
       frameRate: 30,
@@ -66,8 +92,16 @@ const sanitizeProperties = (properties) => {
       videoContentHint: '',
       videoSource: 'camera',
       scalableScreenshare: false,
-    };
+  };
   }
+
+  if (typeof properties.audioFallbackEnabled !== 'undefined') {
+    console.log('audioFallbackEnabled is deprecated -- use audioFallback.');
+    if (properties.audioFallback) {
+      delete properties.audioFallbackEnabled;
+    }
+  }
+
   return {
     videoTrack: sanitizeBooleanProperty(properties.videoTrack),
     audioTrack: sanitizeBooleanProperty(properties.audioTrack),
@@ -76,7 +110,15 @@ const sanitizeProperties = (properties) => {
     publishCaptions: sanitizeBooleanProperty(properties.publishCaptions),
     name: properties.name ? properties.name : '',
     cameraPosition: sanitizeCameraPosition(properties.cameraPosition),
-    audioFallbackEnabled: sanitizeBooleanProperty(properties.audioFallbackEnabled),
+    publisherAudioFallback: sanitizePublisherAudioFallback(
+      properties.audioFallback,
+      properties.videoSource,
+    ),
+    subscriberAudioFallback: sanitizeSubscriberAudioFallback(
+      properties.audioFallback,
+      properties.audioFallbackEnabled,
+      properties.videoSource,
+    ),
     audioBitrate: sanitizeAudioBitrate(properties.audioBitrate),
     enableDtx: sanitizeBooleanProperty(properties.enableDtx ? properties.enableDtx : false),
     frameRate: sanitizeFrameRate(properties.frameRate),
@@ -101,6 +143,10 @@ const sanitizePublisherEvents = (publisherId, events) => {
       rtcStatsReport: 'rtcStatsReport',
       videoNetworkStats: 'videoStats',
       muteForced: 'muteForced',
+      videoDisabled: 'videoDisabled',
+      videoEnabled: 'videoEnabled',
+      videoDisableWarning: 'videoDisableWarning',
+      videoDisableWarningLifted: 'videoDisableWarningLifted',
     },
     android: {
       streamCreated: 'onStreamCreated',
@@ -111,6 +157,10 @@ const sanitizePublisherEvents = (publisherId, events) => {
       rtcStatsReport: 'onRtcStatsReport',
       videoNetworkStats: 'onVideoStats',
       muteForced: 'onMuteForced',
+      videoDisabled: 'onVideoDisabled',
+      videoEnabled: 'onVideoEnabled',
+      videoDisableWarning: 'onVideoDisableWarning',
+      videoDisableWarningLifted: 'onVideoDisableWarningLifted',
     },
   };
   return reassignEvents('publisher', customEvents, events, publisherId);